Убирате лишние запросы на сервер
