+++ src/components/SavedMovies/SavedMovies.js
@@ -4,45 +4,36 @@ import SearchForm from "../SearchForm/SearchForm";
 import "./SavedMovies.css";
 import MoviesCardList from "../MoviesCardList/MoviesCardList.js";
 import { useEffect, useState } from 'react';
-import { api } from "../../utils/Main.js";
 import { search, shortFilmsFilter } from "../../utils/filterFunctions.js";
 import NotSearchResults from "../NotSearchResults/NotSearchResults.js";
 
-function SavedMovies({ pageUrl }) {
-    const [movies, setMovies] = useState([]);
+function SavedMovies({ pageUrl, savedMovies, deleteMovie, removeMovieErrors }) {
     const [errors, setErrors] = useState('');
     const [searchResults, setSearchResults] = useState([]);
     const [notSearchResults, setNotSearchResults] = useState(false);
     const [displayedMovies, setDisplayedMovies] = useState([]);
     const [isShortFilmsFilter, setIsShortFilmsFilter] = useState(false);
-    const [removeMovieErrors, setRemoveMovieErrors] = useState({ message: '', movieId: '' });
     const [isSearch, setIsSearch] = useState(false);
 
     useEffect(() => {
-        api.getMovies()
-            .then((moviesData) => {
-                setMovies(moviesData);
-                setDisplayedMovies(moviesData);
-            })
-            .catch(() => {
-                setErrors("Произошла ошибка. Подождите немного и попробуйте ещё раз")
-            })
+        setDisplayedMovies(savedMovies);
     }, []);
 
     function handleDeleteMovie(movieId) {
-        api.removeMovie(movieId)
-            .then(() => {
-                if ((movies.length === 1 && movies[0]._id === movieId) ||
-                    (isSearch && searchResults.length === 1 && searchResults[0]._id === movieId)) {
-                    setNotSearchResults(true);
-                }
-                setMovies((state) => state.filter((c) => c._id !== movieId));
-                setSearchResults((state) => state.filter((c) => c._id !== movieId));
-                setDisplayedMovies((state) => state.filter((c) => c._id !== movieId));
-            })
-            .catch(() => {
-                setRemoveMovieErrors({ message: "Произошла ошибка. Пожалуйста, попробуйте позже.", movieId: movieId });
-            });
+        deleteMovie(movieId);
+
+        if ((savedMovies.length === 1 && savedMovies[0]._id === movieId) ||
+            (isSearch && searchResults.length === 1 && searchResults[0]._id === movieId)) {
+            setNotSearchResults(true);
+        }
+        if (errors.length === 0) {
+            const movies = isSearch
+                ? searchResults.filter((c) => c._id !== movieId)
+                : savedMovies.filter((c) => c._id !== movieId);
+            getDisplayedMovies(movies, isShortFilmsFilter);
+            setSearchResults((state) => state.filter((c) => c._id !== movieId));
+            setDisplayedMovies((state) => state.filter((c) => c._id !== movieId));
+        }
     }
 
     function getDisplayedMovies(movies, isFiltered) {
@@ -67,7 +58,7 @@ function SavedMovies({ pageUrl }) {
         }
         setSearchResults([]);
         setNotSearchResults(false);
-        let results = search(movies, searchValue);
+        let results = search(savedMovies, searchValue);
         setIsSearch(true);
         setSearchResults(results)
         getDisplayedMovies(results, isShortFilmsFilter);
@@ -79,7 +70,7 @@ function SavedMovies({ pageUrl }) {
         if (isSearch)
             getDisplayedMovies(searchResults, isChecked);
         else
-            getDisplayedMovies(movies, isChecked);
+            getDisplayedMovies(savedMovies, isChecked);
         setIsShortFilmsFilter(isChecked);
     }
 
@@ -89,7 +80,7 @@ function SavedMovies({ pageUrl }) {
             <main>
                 <SearchForm handleSearch={handleSearch} handleShortFilmsFilter={handleShortFilmsFilter} errors={errors} isShortFilmsFilter={isShortFilmsFilter} />
                 {notSearchResults && <NotSearchResults />}
-                {displayedMovies.length !== 0 && <MoviesCardList movies={displayedMovies} errors={removeMovieErrors} isSavedMoviesPage={true} handleDeleteMovie={handleDeleteMovie} savedMovies={movies} />}
+                {displayedMovies.length !== 0 && <MoviesCardList movies={displayedMovies} errors={removeMovieErrors} isSavedMoviesPage={true} handleDeleteMovie={handleDeleteMovie} savedMovies={savedMovies} />}
             </main>
             <Footer />
         </div>
